requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
anthropic==0.39.0
httpx>=0.23.0,<0.28.0
python-dotenv==1.0.0
